Rename "form" folder "inputs"
